Added default descriptor types.
- Added default descriptor type enums.
- Added new functions for default descriptor types.
- Fixed SelectWriteBufferDescriptor now uses a default descriptor type of uniform buffer--since SelectWriteImageDescriptor defaults to combined image sampler.
- Fixed type/size mismatch by static cast in the renderers.

+++ TinyVulkan/TinyVk_GraphicsRenderers.hpp
@@ -310,7 +310,7 @@
 					if (commandPool->rentQueue[i] != false)
 						commandBuffers.push_back(commandPool->commandBuffers[i]);
 				}
-				submitInfo.commandBufferCount = commandBuffers.size();
+				submitInfo.commandBufferCount = static_cast<uint32_t>(commandBuffers.size());
 				submitInfo.pCommandBuffers = commandBuffers.data();
 
 				if (vkQueueSubmit(graphicsPipeline.graphicsQueue, 1, &submitInfo, renderTarget->imageWaitable) != VK_SUCCESS)
@@ -391,7 +391,7 @@
 					if (commandPools[currentSyncFrame]->rentQueue[i] != false)
 						commandBuffers.push_back(commandPools[currentSyncFrame]->commandBuffers[i]);
 				}
-				submitInfo.commandBufferCount = commandBuffers.size();
+				submitInfo.commandBufferCount = static_cast<uint32_t>(commandBuffers.size());
 				submitInfo.pCommandBuffers = commandBuffers.data();
 
 				VkSemaphore signalSemaphores[] = { renderFinishedSemaphores[currentSyncFrame] };
